// src/app/shared/model/page.model.ts
export interface CompanyAccount {
  sNo?: number;
  isSelected: boolean;
  CompanyName: string;
  Email: string;
  AccountURL: string;
  Plan: string;
  CreatedDate: string; 
  Image: string;
  Status: string;
}